fix failing tests 2
